news_aggregator/requirements.txt
# Nepal News Intelligence - Python Dependencies
# Required packages for automated news collection

# RSS Feed Parsing
feedparser==6.0.10

# HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.2

# Data Processing
pandas==2.0.3
numpy==1.24.3

# NLP & Text Processing
scikit-learn==1.3.0

# Visualization (for dashboard)
plotly==5.17.0
matplotlib==3.7.2
wordcloud==1.9.2

# Database
# sqlite3 is built-in to Python, no install needed

# Utilities
python-dateutil==2.8.2
pytz==2023.3
